backend/Greenhouse/requirements.txt
asgiref==3.8.1
dj-rest-auth==6.0.0
Django==5.0.6
django-cors-headers==3.10.1
django-extensions==3.1.5
django-filter==24.2
django-rest-auth==0.9.5
djangorestframework==3.15.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
logger==1.4
Markdown==3.6
mysqlclient==2.2.4
paho-mqtt==2.1.0
PyJWT==2.8.0
python-dotenv==1.0.1
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
